+++ src/main/java/io/github/lix3nn53/guardiansofadelia/bungeelistener/products/BoostPremium.java
@@ -72,13 +72,13 @@ public enum BoostPremium {
         return 0;
     }
 
-    public double applyTo(double value) {
+    public float applyTo(float value) {
         switch (this) {
             case EXPERIENCE:
             case LOOT:
-                return value * 1.2; // +20% bonus value
+                return (float) (value * 1.2); // +20% bonus value
             case ENCHANT:
-                return value + 0.15; // +15% bonus chance
+                return (float) (value + 0.15); // +15% bonus chance
             case GATHER:
                 return value / 2; // +50 faster gathering
         }
